+++ astro.config.mjs
@@ -3,12 +3,10 @@ import solidJs from "@astrojs/solid-js";
 import { defineConfig } from "astro/config";
 import konamimojisplosion from "astro-konamimojisplosion";
 
-import { remarkReadingTime } from "./remark-reading-time.mjs";
-
 export default defineConfig({
 	integrations: [konamimojisplosion(), solidJs(), prefetch()],
 	markdown: {
-		remarkPlugins: [remarkReadingTime],
+		syntaxHighlight: "prism",
 	},
 	site: "https://joshuakgoldberg.com",
 });
