+++ src/components/CenteredTop.module.css
@@ -1,0 +1,19 @@
+.centeredTop {
+	display: flex;
+	flex-direction: column;
+	gap: 1rem;
+	margin: clamp(1rem, 5vh, 3rem) auto;
+	max-width: var(--widthSlim);
+	text-align: center;
+}
+
+.title {
+	line-height: 0.9;
+	text-wrap: balance;
+}
+
+@media (min-width: 700px) {
+	.top {
+		text-align: center;
+	}
+}
