+++ src/components/ContentEntryImage.module.css
@@ -1,0 +1,24 @@
+.contentEntryImage {
+	height: 3.5rem;
+	width: 3.5rem;
+}
+
+.round {
+	border-radius: 100%;
+}
+
+/* todo: switch to webp or avif or some such... or Astro images? */
+.contentEntryImage[src$=".png"] {
+	image-rendering: pixelated;
+}
+
+@media (min-width: 700px) {
+	.contentEntry:not(.widthsFull) {
+		width: calc(50% - 1.5rem);
+	}
+
+	.contentEntryImage {
+		height: 5rem;
+		width: 5rem;
+	}
+}
