+++ src/components/EntryList.module.css
@@ -4,10 +4,11 @@
 	gap: 3.5rem 3rem;
 	list-style-type: none;
 	padding: 0;
+	margin-bottom: clamp(4rem, 10vh, 7rem);
 }
 
 .subHeading {
-	margin: clamp(4rem, 10vh, 7rem) auto 3rem;
+	margin: 3rem auto;
 	scroll-margin-top: 2rem;
 	text-align: center;
 }
