+++ src/components/Header.astro
@@ -3,7 +3,7 @@ import HeaderLink from "./HeaderLink.astro";
 import { Text } from "./Text";
 
 const links = [
-	["Blog", "https://blog.joshuakgoldberg.com"],
+	["Blog", "/blog"],
 	["Book", "https://learningtypescript.com"],
 	["Projects", "/projects", "Assorted project work also seen on my GitHub."],
 	["Speaking", "/speaking"],
