+++ src/components/Hero.astro
@@ -1,27 +1,41 @@
 ---
 import { Text } from "./Text";
+
+const variants = {
+	centered: "centered",
+	left: "left",
+}
+
+export type HeroVariant = keyof typeof variants;
+
 export interface ColorSchemeImages {
 	dark: string;
 	light: string;
 }
 
-interface Props {
+export interface HeroImage {
+	alt: string;
+	src: string | ColorSchemeImages;
+}
+
+export interface Props {
 	class?: string | undefined;
-	image: string | ColorSchemeImages;
+	image: HeroImage;
 	primary: string;
+	variant: HeroVariant;
 }
 
 const { class: className, image, primary } = Astro.props;
-const images =
-	typeof image === "string" ? { dark: image, light: image } : image;
+const src =
+	typeof image.src === "string" ? { dark: image.src, light: image.src } : image.src;
 ---
 
-<div class:list={["hero", className]}>
+<div class:list={["hero", className, variants[Astro.props.variant]]}>
 	<div class="half first">
 		<picture>
-			<source srcset={images.dark} media="(prefers-color-scheme: dark)" />
-			<source srcset={images.light} media="(prefers-color-scheme: light)" />
-			<img alt="" src={images.light} />
+			<source srcset={src.dark} media="(prefers-color-scheme: dark)" />
+			<source srcset={src.light} media="(prefers-color-scheme: light)" />
+			<img alt={image.alt} src={src.light} />
 		</picture>
 	</div>
 	<div class="half second">
@@ -34,6 +48,7 @@ const images =
 
 <style>
 	.hero {
+		--imageSize: 10rem;
 		align-items: center;
 		display: flex;
 		flex-direction: column;
@@ -41,32 +56,42 @@ const images =
 		margin: 1rem auto 4rem;
 		padding: 0 1rem;
 		position: relative;
-		width: var(--widthFull);
-		max-width: 100%;
 		text-align: center;
+	}
 
-		--imageSize: 10rem;
+	.hero.centered {
+		--h1LineHeight: 0.75;
+		max-width: 100%;
+		margin: 1rem auto 4rem;
+		width: var(--widthFull);
+	}
+
+	.hero.left {
+		--h1LineHeight: 0.9;
+		margin: 2rem auto 0;
+		max-width: var(--widthSlim);
 	}
 
 	h1 {
 		color: var(--colorForegroundExtra);
-		line-height: 0.75;
+		line-height: var(--h1LineHeight);
 		text-shadow: var(--textShadow);
 	}
 
 	img {
 		user-select: none;
+		height: var(--imageSize);
 		width: var(--imageSize);
 	}
 
 	@media (min-width: 700px) {
 		.hero {
+			flex-direction: row;
 			gap: 2.5rem;
 			height: 35vh;
-			min-height: calc(var(--imageSize) + 1rem);
-			max-height: 20rem;
-			flex-direction: row;
 			margin: clamp(1rem, 5vh, 7rem) 0 clamp(2rem, 7vh, 10rem);
+			max-height: 20rem;
+			min-height: calc(var(--imageSize) + 1rem);
 			padding: 0;
 			text-align: left;
 		}
@@ -79,6 +104,9 @@ const images =
 			display: flex;
 			justify-content: flex-end;
 			max-height: initial;
+		}
+
+		.centered .half.first {
 			width: 45%;
 		}
 
@@ -87,6 +115,9 @@ const images =
 			flex-direction: column;
 			align-items: flex-start;
 			justify-content: center;
+		}
+
+		.centered .half.second {
 			width: 55%;
 		}
 
@@ -95,12 +126,16 @@ const images =
 		}
 
 		img {
+			aspect-ratio: 1;
 			border-radius: 100%;
+			box-shadow: 0 7px 7px 3px var(--colorSubtle);
+		}
+
+		.centered img {
 			height: 256px;
 			max-height: 100%;
 			max-width: 100%;
 			width: 256px;
-			box-shadow: 0 7px 7px 3px var(--colorSubtle);
 		}
 	}
 </style>
