--- src/components/ProjectEntry.module.css
@@ -1,7 +1,0 @@
-.subList {
-	display: flex;
-	flex-direction: column;
-	gap: 0.5rem;
-	list-style-type: none;
-	padding-left: 0;
-}
