--- src/components/SpeakingEntry.module.css
@@ -1,8 +1,0 @@
-.event {
-	margin: 0.35rem 0;
-}
-
-.links {
-	display: flex;
-	gap: 0.5rem;
-}
