+++ src/components/blog/DateAndMinutes.module.css
@@ -1,0 +1,8 @@
+.between {
+	display: inline-block;
+	background: var(--colorForeground);
+	border-radius: 100%;
+	width: 0.3em;
+	height: 0.3em;
+	margin: 0 0.75em 0.15em;
+}
