+++ src/components/speaking/ProjectEntry.module.css
@@ -1,0 +1,7 @@
+.subList {
+	display: flex;
+	flex-direction: column;
+	gap: 0.5rem;
+	list-style-type: none;
+	padding-left: 0;
+}
