+++ src/layouts/base.css
@@ -22,6 +22,22 @@
 		--textShadow: 3px 3px 1px var(--colorSubtle);
 
 		--widthFull: 1000px;
+		--widthSlim: 49rem;
+
+		/* Note: some of these are unused in Prism, but would be in Shiki. */
+		--astro-code-color-background: transparent;
+		--astro-code-color-deleted: #903838;
+		--astro-code-color-inserted: #0e601c;
+		--astro-code-color-text: #232642;
+		--astro-code-token-comment: #139e4b;
+		--astro-code-token-constant: #001080;
+		--astro-code-token-function: #466da3;
+		--astro-code-token-keyword: #0058cb;
+		--astro-code-token-link: #e7ffe7;
+		--astro-code-token-parameter: #588591;
+		--astro-code-token-punctuation: #0f90a1;
+		--astro-code-token-string-expression: #408aa7;
+		--astro-code-token-string: #674cab;
 	}
 
 	html {
@@ -60,12 +76,24 @@
 
 	@media (prefers-color-scheme: dark) {
 		:root {
-			--colorBackgroundBottom: #141117;
+			--colorBackgroundBottom: #111117;
 			--colorBackgroundTop: #071428;
-			--colorEmphasized: #78b4f0;
+			--colorEmphasized: #77bbff;
 			--colorForeground: #dae7f3;
 			--colorForegroundExtra: #e5edf3;
 			--colorSubtle: #052438;
+
+			/* Note: some of these are unused in Prism, but would be in Shiki. */
+			--astro-code-color-text: #d3d6f0;
+			--astro-code-token-comment: #85cab1;
+			--astro-code-token-constant: #d4d8f4;
+			--astro-code-token-function: #85b8ff;
+			--astro-code-token-keyword: #b0cff7;
+			--astro-code-token-link: #bde3bd;
+			--astro-code-token-parameter: #8dc1d0;
+			--astro-code-token-punctuation: #55e1f3;
+			--astro-code-token-string-expression: #82c1da;
+			--astro-code-token-string: #b3baf2;
 		}
 	}
