+++ src/pages/blog/slug.module.css
@@ -1,0 +1,72 @@
+.top {
+	margin: clamp(1rem, 5vh, 3rem) 0;
+}
+
+.article {
+	display: flex;
+	flex-direction: column;
+	font-size: 1.35rem;
+	gap: 1rem;
+	font-weight: 300;
+	line-height: 1.5;
+	margin: 3rem auto 2rem;
+	max-width: calc(var(--widthSlim) + 4rem);
+	padding: 0 2rem;
+}
+
+.article a {
+	text-decoration: underline;
+}
+
+.article h2 {
+	margin-top: 2rem;
+}
+
+.article h3 {
+	margin-top: 1rem;
+}
+
+.article code {
+	font-size: 0.9em;
+}
+
+.article pre {
+	border: 1px solid var(--colorSubtle);
+	font-size: 0.9em;
+	padding: 1rem 1.5rem;
+	margin: 0;
+	tab-size: 4;
+}
+
+.article blockquote {
+	--blockBorderWidth: 6px;
+	font-style: italic;
+	margin-left: 0;
+	padding-left: calc(2rem - var(--blockBorderWidth));
+	position: relative;
+}
+
+.article blockquote::before {
+	content: "";
+	display: block;
+	position: absolute;
+	width: var(--blockBorderWidth);
+	height: 100%;
+	left: 0;
+	top: 0;
+	background: linear-gradient(
+		to bottom,
+		var(--colorEmphasized),
+		var(--colorForeground)
+	);
+}
+
+.article ul {
+	margin: 0;
+}
+
+@media (min-width: 700px) {
+	.top {
+		text-align: center;
+	}
+}
