+++ src/pages/index.astro
@@ -2,11 +2,11 @@
 import BodyContents from "../components/BodyContents.astro";
 import Socials from "../components/Socials.astro";
 import { Text } from "../components/Text";
-import BaseLayout from "../layouts/BaseLayout.astro";
+import HeroLayout from "../layouts/HeroLayout.astro";
 ---
 
 <!-- image="images/josh.png" -->
-<BaseLayout
+<HeroLayout
 	description="I'm an open source maintainer in the TypeScript ecosystem."
 	primary="Hi! I'm Josh."
 	secondary="I'm an open source maintainer in the TypeScript ecosystem."
@@ -25,7 +25,7 @@ import BaseLayout from "../layouts/BaseLayout.astro";
 			an international conference speaker, a mentor, and a live code streamer.
 		</Text>
 	</BodyContents>
-</BaseLayout>
+</HeroLayout>
 
 <script>
 	// Preserves legacy linking behavior from my old static HTML site.
