+++ src/pages/speaking.astro
@@ -4,13 +4,14 @@ import { getCollection } from "astro:content";
 import { Anchor } from "../components/Anchor";
 import BodyContents from "../components/BodyContents.astro";
 import HeroTertiary from "../components/HeroTertiary.astro";
-import { SpeakingsGrouped } from "../components/SpeakingsGrouped";
-import BaseLayout from "../layouts/BaseLayout.astro";
+import { SpeakingsGrouped } from "../components/speaking/SpeakingsGrouped";
+import HeroLayout from "../layouts/HeroLayout.astro";
 
 const allSpeakings = await getCollection("speaking");
 ---
 
-<BaseLayout
+<HeroLayout
+	alt="Josh looking up seriously wearing a dark blue shirt with a small pineapples and flamingos pattern"
 	description="I talk about static analysis, accessibility, and team culture."
 	primary="Speaking"
 	secondary="I talk about static analysis, accessibility, and team culture."
@@ -31,4 +32,4 @@ const allSpeakings = await getCollection("speaking");
 			speakings={allSpeakings.sort((a, b) => +b.data.date - +a.data.date)}
 		/>
 	</BodyContents>
-</BaseLayout>
+</HeroLayout>
